feat(external): from scratch
