+++ src/no-framework/external-module-mock.js
@@ -1,24 +1,8 @@
-const fn = (impl = () => {}) => {
-  const mockFn = (...args) => {
-    mockFn.mock.calls.push(args);
-    return impl(...args);
-  };
-  mockFn.mock = { calls: [] };
-  mockFn.mockImplementation = newImpl => (impl = newImpl);
-
-  return mockFn;
-};
-
+require('../__no-framework-mocks__/utils');
 const utilsPath = require.resolve('../utils');
+const mockUtilsPath = require.resolve('../__no-framework-mocks__/utils');
+require.cache[utilsPath] = require.cache[mockUtilsPath];
 
-require.cache[utilsPath] = {
-  id: utilsPath,
-  fileName: utilsPath,
-  loaded: true,
-  exports: {
-    getWinner: fn((p1, p2) => p1),
-  },
-};
 const assert = require('assert');
 const thumbWar = require('../thumb-war');
 const utils = require('../utils');
